Extract FormData writer
